Changed cast to avoid error in GCC 4.0. The 'pointer' argument is really an integer which will fit easily into 32 bits.

+++ gtk/SciTEGTK.cxx
@@ -2214,7 +2214,7 @@ gint SciTEGTK::QuitSignal(GtkWidget *, GdkEventAny *, SciTEGTK *scitew) {
 }
 
 void SciTEGTK::ButtonSignal(GtkWidget *, gpointer data) {
-	instance->Command((guint)data);
+	instance->Command((guint)(long)data);
 }
 
 void SciTEGTK::MenuSignal(SciTEGTK *scitew, guint action, GtkWidget *) {
